feat: 모임조회 헤더 레이아웃 추가
